LayoutTests/imported/w3c/web-platform-tests/css/css-shapes/shape-outside/shape-image/gradients/shape-outside-radial-gradient-003-expected.txt

CONSOLE MESSAGE: diff: 1.984375
CONSOLE MESSAGE: diff: 3.8125
CONSOLE MESSAGE: diff: 8
CONSOLE MESSAGE: diff: 25
The test passes if you see green boxes following the contour of a circle. There should be no red.

x
x
x
x
x
x
x
x
x
x
x
x

FAIL CSS Test: Left float with radial gradient shape, shape margin, and shape-image-threshold assert_true: Lines positioned properly around the shape. expected true got false
